+++ src/main/java/com/example/carsharing/repository/CarRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface CarRepository extends JpaRepository<Car,Integer> {
+public interface CarRepository extends JpaRepository<Car,Long> {
     List<Car> findByIsRentedTrue();
     List<Car> findByIsRentedFalse();
-    Optional<Car> findByIdAndIsRentedTrue(Integer id);
+    Optional<Car> findByIdAndIsRentedTrue(Long id);
 
 }
